* unused code removed * makeDataSelector.ts moved to the store folder
