+++ src/store/makeDataSelector.ts
@@ -1,6 +1,6 @@
 import { createSelector } from 'reselect';
 
-import { RootState } from '../store';
+import { RootState } from '.';
 
 type EntityTypes = Pick<RootState, 'user'>;
